requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
distro==1.9.0
Django==5.1.6
django-ckeditor==6.7.2
django-cors-headers==4.3.1
django-filter==24.3
django-jazzmin==3.0.1
django-js-asset==3.0.1
django-meta==2.4.1
django-robots==6.1
djangorestframework==3.14.0
drf-yasg==1.21.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.8.2
openai==1.61.1
packaging==24.2
pillow==10.2.0
pydantic==2.10.6
pydantic_core==2.27.2
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.3
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
